only remove it if it exists
git-svn-id: https://svn.sixfeetup.com/svn/private/claytron/dotfiles/trunk@10726 7abc7aa1-7e52-db11-9331-00508bd72530

+++ create_links.sh
@@ -37,8 +37,10 @@ for dotfile in "${dotfiles[@]}"
             to_create=$HOME/$dotfile
         fi
         if [[ $remove == "unlink" ]]; then
-            rm $to_create
-            echo "Unlinked $to_create"
+            if [ -e $to_create ]; then
+                rm $to_create
+                echo "Unlinked $to_create"
+            fi
         else
             # symlink the conf file
             if [ ! -e $to_create ]; then
